Removed window module - refactored, see below
Added mainwindow module
- Moved functionality from main.rs to separate class, extending libadwaita::ApplicationWindow
- All button functionality now works as menu items in main window
- Keeps track of opening app settings and nvidia-settings (nvidia-settings currently doesn't work)
- Creates settings window as separate app (shouldn't do this! need to store APP_ID as a custom property OR fetch application)
- Focus doesn't currently work as intended (doesn't refocus on settings close)

Added settingswindow module
- For managing app settings
- placeholder for now

Updated main.rs
- Switched to libadwaita Application class and libadwaita gio wrapper
- Renamed APP_ID to "com.gtk_d.NvidiaExtensionRust"
- Moved window into it's own class (massively simplifying main.rs)
