+++ src/main.rs
@@ -19,22 +19,16 @@
  */
 
 // Modules
-mod custom_button;
-//use custom_button::CustomButton;
 mod subprocess;
 mod processor;
-mod property;
-use property::Property;
 mod formatter;
-use formatter::Formatter;
+mod property;
 mod provider;
-use provider::Provider;
-mod window;
-use window::Window;
+mod mainwindow;
+use mainwindow::MainWindow;
+mod settingswindow;
 
 // Imports
-use processor::Processor;
-use std::ffi::OsStr;
 //use std::env;
 //use std::path::Path;
 //use libappindicator::{
@@ -42,26 +36,27 @@ use std::ffi::OsStr;
 //};
 
 use adwaita::prelude::*;
+//use gtk::Application;
 use adwaita::{
-    /* Libraries */ gio, /* Application */ Application, ApplicationWindow,
-    SplitButton
+    /* Libraries */ gio, /* Application */ Application,
+//    SplitButton
 };
-use gio::Menu;
+use gio::{resources_register_include, Settings};
 
 // Constants
-const APP_ID: &str = "org.gtk_d.NvidiaExtensionRust";
+const APP_ID: &str = "com.gtk_d.NvidiaExtensionRust";
 
 // Main Function
 fn main() {
     // Resources
-    gio::resources_register_include!("nvidiaextensionrust.gresource")
+    resources_register_include!("nvidiaextensionrust.gresource")
     .expect("Failed to register resources.");
 
     // Intialise GTK
     gtk::init().expect("Failed to initialise gtk");
 
     // Create a new application
-    let app = Application::builder().application_id(APP_ID).build();
+    let app: Application = Application::builder().application_id(APP_ID).build();
 
     // Connect to "activate" signal of `app`
     //app.connect_startup(setup_shortcuts);
@@ -84,403 +79,10 @@ fn setup_shortcuts(app: &Application) {
 // Build Function
 fn build_ui(app: &Application) {
     // Create a new custom window and show it
-    let window = Window::new(app);
+    let window: MainWindow = MainWindow::new(app);
     window.show();
 
-
-
     /*
-    // Button Child 1: exec_check (subprocess) launch nvidia-settings
-    let button1: SplitButton = SplitButton::builder()
-        .label("Open Settings")
-        .margin_top(12)
-        .margin_bottom(12)
-        .margin_start(12)
-        .margin_end(12)
-        .build();
-    // Connect to "clicked" signal of `button`
-    button1.connect_clicked(move |_| {
-        match subprocess::exec_check(&[OsStr::new("nvidia-settings")], None::<&gio::Cancellable>) {
-            Ok(_x) => println!("Opening the Nvidia Settings app.."),
-            Err(y) => println!(
-                "An error occured while opening the Nvidia Settings app: {}",
-                y.message()
-            ),
-        };
-    });
-    // Button Child 2: exec_communicate (subprocess) ask for GPU data
-    let button2: SplitButton = SplitButton::builder()
-        .label("Get GPU Names (SubProcess)")
-        .margin_top(12)
-        .margin_bottom(12)
-        .margin_start(12)
-        .margin_end(12)
-        .build();
-    // Connect to "clicked" signal of `button`
-    button2.connect_clicked(move |_| {
-        match subprocess::exec_communicate(
-            &[
-                OsStr::new("nvidia-settings"),
-                OsStr::new("-q"),
-                OsStr::new("GpuUUID"),
-                OsStr::new("-t"),
-            ],
-            None::<&gio::Cancellable>,
-        ) {
-            Ok(return_val) => match return_val {
-                (None, None) => println!("no stdout or stderr, something went really wrong..."),
-                (None, Some(stderr_buffer)) => match std::str::from_utf8(&stderr_buffer) {
-                    Ok(stderr_buffer_contents) => {
-                        println!("Process failed with error: {}", stderr_buffer_contents)
-                    }
-                    Err(err) => panic!("{}", err),
-                },
-                (Some(stdout_buffer), None) => match std::str::from_utf8(&stdout_buffer) {
-                    Ok(stdout_buffer_contents) => {
-                        println!("Process suceeded, returning: {}", stdout_buffer_contents)
-                    }
-                    Err(err) => panic!("{}", err),
-                },
-                (Some(stdout_buffer), Some(stderr_buffer)) => {
-                    match std::str::from_utf8(&stdout_buffer) {
-                        Ok(stdout_buffer_contents) => match std::str::from_utf8(&stderr_buffer) {
-                            Ok(stderr_buffer_contents) => println!(
-                                "Process suceeded, returning: {} but with error: {}",
-                                stdout_buffer_contents, stderr_buffer_contents
-                            ),
-                            Err(err) => panic!("{}", err),
-                        },
-                        Err(err) => panic!("{}", err),
-                    }
-                }
-            },
-            Err(err) => println!("something went wrong: {}", err),
-        };
-    });
-    // Button Child 3: Processor ask for GPU data
-    let button3: SplitButton = SplitButton::builder()
-        .label("Get GPU Names (Processor)")
-        .margin_top(12)
-        .margin_bottom(12)
-        .margin_start(12)
-        .margin_end(12)
-        .build();
-    // Connect to "clicked" signal of `button`
-    button3.connect_clicked(move |_| {
-        let p: Processor = Processor::new("nvidia-settings", "-q GpuUUID -t");
-
-        //NOTE: Leaving this here for future use..
-        //p.set_property("base-call", "nvidia-settings");
-        //p.set_property("call", "nvidia-settings");
-        //p.set_property("tail-call", "t");
-
-        match p.process() {
-            Ok(output) => match output {
-                Some(valid_output) => println!("Process suceeded, returning: `{}`", valid_output),
-                None => println!("Process encountered an unknown error.."),
-            },
-            Err(err) => println!("Process encountered an error, returning: `{}`", err),
-        }
-    });
-    // Button Child 4: property test
-    let button4: SplitButton = SplitButton::builder()
-        .label("Test Formatter & Property")
-        .margin_top(12)
-        .margin_bottom(12)
-        .margin_start(12)
-        .margin_end(12)
-        .build();
-    // Connect to "clicked" signal of `button`
-    button4.connect_clicked(move |_| {
-        // GENERIC
-        let proc: Processor = Processor::new("nvidia-settings", "-q GpuUUID -t");
-        let form: Formatter = Formatter::new();
-        let p: Property = Property::new(&proc, "", "", &form, &1);
-
-        let vecc: Vec<Vec<String>> = vec![
-            vec!["1.68".to_string(), "2.01".to_string()],
-            vec!["3.83".to_string(), "4.22".to_string()],
-        ];
-        match p.parse(vecc, |input: Vec<String>| {
-            Some(input.get(0).unwrap().to_string())
-        }) {
-            Some(results) => {
-                println!("size: {}", results.len());
-                for item in results {
-                    println!("item: {}", item);
-                }
-            }
-            None => println!("Something's gone really wrong!"),
-        }
-
-        // PERCENT
-        let proc: Processor = Processor::new("nvidia-settings", "-q GpuUUID -t");
-        let form: Formatter = Formatter::new();
-        let p: Property = Property::new(&proc, "", "", &form, &1);
-        let vecc: Vec<Vec<String>> = vec![
-            vec!["1.68".to_string(), "2.01".to_string()],
-            vec!["3.83".to_string(), "4.22".to_string()],
-        ];
-        match p.parse(vecc, |input: Vec<String>| {
-            // Grab input
-            let mut output: String = input.get(0).unwrap().to_string();
-
-            // Apply formatting
-            output.push('%');
-
-            // Return result
-            Some(output)
-        }) {
-            Some(results) => {
-                println!("size: {}", results.len());
-                for item in results {
-                    println!("item: {}", item);
-                }
-            }
-            None => println!("Something's gone really wrong when formatting GENERIC info"),
-        }
-
-        // POWER
-        let proc: Processor = Processor::new("nvidia-settings", "-q GpuUUID -t");
-        let form: Formatter = Formatter::new();
-        let p: Property = Property::new(&proc, "", "", &form, &1);
-        let vecc: Vec<Vec<String>> = vec![
-            vec!["1.68".to_string(), "2.01".to_string()],
-            vec!["3.83".to_string(), "4.22".to_string()],
-        ];
-        match p.parse(vecc, |input: Vec<String>| {
-            // Grab input
-            let input_str: String = input.get(0).unwrap().to_string();
-
-            // Convert to float
-            match input_str.parse::<f64>() {
-                Ok(parsed_value) => {
-                    // Round down to nearest integer
-                    let rounded_value: f64 = parsed_value.floor();
-
-                    // Convert to string
-                    let mut output: String = rounded_value.to_string();
-
-                    // Apply formatting
-                    output.push('W');
-
-                    // Return result
-                    Some(output)
-                }
-                Err(_) => {
-                    //this should catch "" etc
-                    println!("Not a valid number");
-
-                    None
-                }
-            }
-        }) {
-            Some(results) => {
-                println!("size: {}", results.len());
-                for item in results {
-                    println!("item: {}", item);
-                }
-            }
-            None => println!("Something's gone really wrong when formatting POWER info"),
-        }
-
-        // MEMORY
-        let proc: Processor = Processor::new("nvidia-settings", "-q GpuUUID -t");
-        let form: Formatter = Formatter::new();
-        let p: Property = Property::new(&proc, "", "", &form, &1);
-        let vecc: Vec<Vec<String>> = vec![
-            vec!["1.68".to_string(), "2.01".to_string()],
-            vec!["3.83".to_string(), "4.22".to_string()],
-        ];
-        match p.parse(vecc, |input: Vec<String>| {
-            // Grab input
-            let current: String = input.get(0).unwrap().to_string();
-            let max: String = input.get(1).unwrap().to_string();
-
-            // Convert to float
-            match current.parse::<f64>() {
-                Ok(parsed_current) => {
-                    match max.parse::<f64>() {
-                        Ok(parsed_max) => {
-                            // Calculate total memory usage
-                            let usage: f64 = (parsed_current / parsed_max) * 100.0;
-
-                            // Round down to nearest integer
-                            let rounded_value: f64 = usage.floor();
-
-                            // Convert to string
-                            let mut output: String = rounded_value.to_string();
-
-                            // Apply formatting
-                            output.push('%');
-
-                            // Return result
-                            Some(output)
-                        }
-                        Err(_) => {
-                            //this should catch "" etc
-                            println!("Not a valid number");
-
-                            None
-                        }
-                    }
-                }
-                Err(_) => {
-                    //this should catch "" etc
-                    //TODO: fix this!
-                    println!("Not a valid number");
-
-                    None
-                }
-            }
-        }) {
-            Some(results) => {
-                println!("size: {}", results.len());
-                for item in results {
-                    println!("item: {}", item);
-                }
-            }
-            None => println!("Something's gone really wrong when formatting MEMORY info"),
-        }
-
-        // TEMPERATURE
-        let proc: Processor = Processor::new("nvidia-settings", "-q GpuUUID -t");
-        let form: Formatter = Formatter::new();
-        let p: Property = Property::new(&proc, "", "", &form, &1);
-        let vecc: Vec<Vec<String>> = vec![
-            vec!["1.68".to_string(), "2.01".to_string()],
-            vec!["3.83".to_string(), "4.22".to_string()],
-        ];
-        match p.parse(vecc, |input: Vec<String>| {
-            // Grab input
-            let mut output: String = input.get(0).unwrap().to_string();
-
-            //TODO: needs moved to settings
-            #[derive(Debug, PartialEq, Eq)]
-            enum TemperatureUnit {
-                CELCIUS = 0,
-                FAHRENHEIT = 1,
-            }
-            let current_unit: TemperatureUnit = TemperatureUnit::CELCIUS;
-            //let current_unit: TemperatureUnit = TemperatureUnit::FAHRENHEIT;
-
-            // Apply formatting
-            if current_unit == TemperatureUnit::CELCIUS {
-                // Apply temperature unit
-                output.push(char::from_u32(0x00B0).unwrap());
-                output.push('C');
-            } else if current_unit == TemperatureUnit::FAHRENHEIT {
-                match output.parse::<f64>() {
-                    Ok(temp) => {
-                        // Convert to fahrenheit
-                        let fahrenheit_temp: f64 = temp * 9.0 / 5.0 + 32.0;
-
-                        // Round down to nearest integer
-                        let rounded_value: f64 = fahrenheit_temp.floor();
-
-                        // Convert to string
-                        let mut f_output: String = rounded_value.to_string();
-
-                        // Apply temperature unit
-                        f_output.push(char::from_u32(0x00B0).unwrap());
-                        f_output.push('F');
-
-                        // Return result
-                        Some(f_output)
-                    }
-                    Err(_) => {
-                        //this should catch "" etc
-                        println!("Not a valid number");
-
-                        None
-                    }
-                };
-            }
-
-            // Return result
-            Some(output)
-        }) {
-            Some(results) => {
-                println!("size: {}", results.len());
-                for item in results {
-                    println!("item: {}", item);
-                }
-            }
-            None => println!("Something's gone really wrong when formatting TEMPERATURE info"),
-        }
-    });
-    // Button Child 5: provider test
-    let button5: SplitButton = SplitButton::builder()
-        .label("Test Providers")
-        .margin_top(12)
-        .margin_bottom(12)
-        .margin_start(12)
-        .margin_end(12)
-        .build();
-    // Connect to "clicked" signal of `button`
-    button5.connect_clicked(move |_| {
-        // SETTINGS
-        let settings_prov: Provider = Provider::new(|| {
-            let proc: Processor = Processor::new("nvidia-settings", "-q GpuUUID -t");
-            let form: Formatter = Formatter::new();
-            let gpu_count: i32 = 1;
-
-            vec![
-                Property::new(&proc, "utilization.gpu", "", &form, &gpu_count),
-                Property::new(&proc, "temperature.gpu", "", &form, &gpu_count),
-                Property::new(&proc, "memory.used,memory.total", "", &form, &gpu_count),
-                Property::new(&proc, "fan.speed", "", &form, &gpu_count),
-            ]
-        });
-
-        // SMI
-        let smi_prov: Provider = Provider::new(|| {
-            let proc: Processor = Processor::new("nvidia-smi", "--query-gpu=gpu_name --format=csv,noheader");
-            let form: Formatter = Formatter::new();
-            let gpu_count: i32 = 1;
-
-            vec![
-                Property::new(&proc, "utilization.gpu", "", &form, &gpu_count),
-                Property::new(&proc, "temperature.gpu", "", &form, &gpu_count),
-                Property::new(&proc, "memory.used,memory.total", "", &form, &gpu_count),
-                Property::new(&proc, "fan.speed", "", &form, &gpu_count),
-                Property::new(&proc, "power.draw", "", &form, &gpu_count),
-            ]
-        });
-
-        // SETTINGS & SMI
-        let both_prov: Provider = Provider::new(|| {
-            let settings_proc: Processor = Processor::new("nvidia-smi", "--query-gpu=gpu_name --format=csv,noheader");
-            let smi_proc: Processor = Processor::new("nvidia-smi", "--query-gpu=gpu_name --format=csv,noheader");
-            let form: Formatter = Formatter::new();
-            let gpu_count: i32 = 1;
-
-            vec![
-                Property::new(&settings_proc, "utilization.gpu", "", &form, &gpu_count),
-                Property::new(&settings_proc, "temperature.gpu", "", &form, &gpu_count),
-                Property::new(&settings_proc, "memory.used,memory.total", "", &form, &gpu_count),
-                Property::new(&settings_proc, "fan.speed", "", &form, &gpu_count),
-                Property::new(&smi_proc, "power.draw", "", &form, &gpu_count),
-            ]
-        });
-
-        // OPTIMUS
-        let optimus_prov: Provider = Provider::new(|| {
-            let proc: Processor = Processor::new("optirun", "nvidia-smi --query-gpu=gpu_name --format=csv,noheader");
-            let form: Formatter = Formatter::new();
-            let gpu_count: i32 = 1;
-
-            vec![
-                Property::new(&proc, "utilization.gpu", "", &form, &gpu_count),
-                Property::new(&proc, "temperature.gpu", "", &form, &gpu_count),
-                Property::new(&proc, "memory.used,memory.total", "", &form, &gpu_count),
-                Property::new(&proc, "fan.speed", "", &form, &gpu_count),
-                Property::new(&proc, "power.draw", "", &form, &gpu_count),
-            ]
-        });
-
-    });
-
     // Menu Child
     let menu: Menu = Menu::new();
     let item: Menu = Menu::new();
