+++ src/mainwindow/imp.rs
@@ -1,0 +1,198 @@
+// SPDX-FileCopyrightText: 2022 Deren Vural
+// SPDX-License-Identifier: GPL-3.0-or-later
+
+/*
+ * Name:
+ * imp.rs
+ *
+ * Description:
+ * Implementation of our custom GObject class (MainWindow)
+ *
+ * Made:
+ * 09/10/2022
+ *
+ * Made by:
+ * Deren Vural
+ *
+ * Notes:
+ *
+ */
+
+// Imports
+use adwaita::subclass::prelude::*;
+use gtk::subclass::prelude::*;
+
+use gio::Settings;
+use glib::signal::Inhibit;
+use glib::subclass::InitializingObject;
+
+use adwaita::prelude::*;
+use gtk::{gio, glib, CompositeTemplate};//, Entry, ListBox, TemplateChild};
+use glib::{once_cell::sync::OnceCell, Value, once_cell::sync::Lazy, ParamSpec};
+
+// Modules
+//use crate::utils::data_path;
+
+#[derive(CompositeTemplate, Default)]
+#[template(resource = "/main-window.ui")]
+pub struct MainWindow {
+    pub settings: OnceCell<Settings>,
+}
+
+// The central trait for subclassing a GObject
+#[glib::object_subclass]
+impl ObjectSubclass for MainWindow {
+    // `NAME` needs to match `class` attribute of template
+    const NAME: &'static str = "NvidiaExtensionMainWindow";
+    type Type = super::MainWindow;
+    type ParentType = gtk::ApplicationWindow;
+
+    fn class_init(klass: &mut Self::Class) {
+        klass.bind_template();
+    }
+
+    fn instance_init(obj: &InitializingObject<Self>) {
+        obj.init_template();
+    }
+}
+
+/*
+ * Trait Name:
+ * ObjectImpl
+ *
+ * Description:
+ * Trait shared by all GObjects
+ *
+ * Made:
+ * 09/10/2022
+ *
+ * Made by:
+ * Deren Vural
+ *
+ * Notes:
+ *
+ */
+impl ObjectImpl for MainWindow {
+    fn constructed(&self, obj: &Self::Type) {
+        // Call "constructed" on parent
+        self.parent_constructed(obj);
+
+        // Setup
+        obj.setup_settings();
+        //obj.setup_tasks();
+        obj.restore_data();
+        obj.setup_callbacks();
+        obj.setup_actions();
+    }
+}
+
+/*
+ * Trait Name:
+ * WidgetImpl
+ *
+ * Description:
+ * Trait shared by all widgets
+ *
+ * Made:
+ * 09/10/2022
+ *
+ * Made by:
+ * Deren Vural
+ *
+ * Notes:
+ *
+ */
+impl WidgetImpl for MainWindow {}
+
+/*
+ * Trait Name:
+ * WindowImpl
+ *
+ * Description:
+ * Trait shared by all Window's
+ *
+ * Made:
+ * 09/10/2022
+ *
+ * Made by:
+ * Deren Vural
+ *
+ * Notes:
+ *
+ */
+impl WindowImpl for MainWindow {
+    fn close_request(&self, window: &Self::Type) -> Inhibit {
+        /*
+        // Store task data in vector
+        let backup_data: Vec<TaskData> = window
+            .tasks()
+            .snapshot()
+            .iter()
+            .filter_map(Cast::downcast_ref::<TaskObject>)
+            .map(TaskObject::task_data)
+            .collect();
+
+        // Save state to file
+        let file = File::create(data_path()).expect("Could not create json file.");
+        serde_json::to_writer(file, &backup_data)
+            .expect("Could not write data to json file");
+
+        */
+        // Pass close request on to the parent
+        self.parent_close_request(window)
+    }
+}
+
+/*
+ * Trait Name:
+ * AdwWindowImpl
+ *
+ * Description:
+ * Trait shared by all AdwWindow's
+ *
+ * Made:
+ * 09/10/2022
+ *
+ * Made by:
+ * Deren Vural
+ *
+ * Notes:
+ *
+ */
+impl AdwWindowImpl for MainWindow {}
+
+/*
+ * Trait Name:
+ * ApplicationWindowImpl
+ *
+ * Description:
+ * Trait shared by all ApplicationWindow's
+ *
+ * Made:
+ * 09/10/2022
+ *
+ * Made by:
+ * Deren Vural
+ *
+ * Notes:
+ *
+ */
+impl ApplicationWindowImpl for MainWindow {}
+
+/*
+ * Trait Name:
+ * AdwApplicationWindowImpl
+ *
+ * Description:
+ * Trait shared by all AdwApplicationWindow's
+ *
+ * Made:
+ * 09/10/2022
+ *
+ * Made by:
+ * Deren Vural
+ *
+ * Notes:
+ *
+ */
+impl AdwApplicationWindowImpl for MainWindow {}
